+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## v0.2.0
+
+### VSCode/RStudio/tmTheme
+
+-   Improve token definition and add colors.
+
 ## v0.1.3
 
--   Also available in <https://open-vsx.org/> 
+-   Also available in <https://open-vsx.org/>
     [[Link]](https://open-vsx.org/extension/dieghernan/skeletor-syntax).
 
 ## v0.1.2